contour tracker
The challenging part of this program was to find a way for the computer to distinguish two or more clusters. These clusters are curves and strands. By this we mean that their thickness is much smaller than its length. Moreover, and they can take many forms. They also evolve in time, which means we need to apply the algorithm each time step. The result is achieved. We can distinguish one strand from the other, but as time advances, we are unable track (label) the cluster in time.  The data that the program can read is set to be format the yolohtli repository.